refactor: image to shared
