+++ src/app/components/course-card/course-card.tsx
@@ -2,7 +2,7 @@ import { Link } from 'react-router-dom';
 import { type Course } from '@/app/types';
 import micIcon from '@/assets/icons/mic.svg';
 import noteIcon from '@/assets/icons/note-icon.svg';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 import './course-card.scss';
