+++ src/features/study-path/components/stage-card/logo-icon/logo-icon.tsx
@@ -1,4 +1,4 @@
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 import './logo-icon.scss';
