+++ src/icons/angular-icon.tsx
@@ -1,5 +1,5 @@
 import angular from '@/assets/svg/angular-icon.svg';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 export const AngularIcon = () => {
   return <Image src={angular} alt="angular icon" />;
