+++ src/icons/aws.tsx
@@ -1,5 +1,5 @@
 import aws from '@/assets/svg/aws.svg';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 export const AwsLogo = () => {
   return <Image src={aws} alt="AWS icon" />;
