+++ src/icons/discord-logo.tsx
@@ -1,5 +1,5 @@
 import discordLogo from '@/assets/svg/discord-logo.svg';
-import Image from '@/features/image/image';
+import Image from '@/shared/image/image';
 
 export const DiscordLogo = () => {
   return <Image src={discordLogo} alt="discord logo" />;
