+++ src/icons/epam.tsx
@@ -1,5 +1,5 @@
 import epam from '@/assets/svg/epam.svg';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 export const EpamLogo = () => {
   return <Image src={epam} alt="epam" />;
