+++ src/icons/external-link-icon.tsx
@@ -1,5 +1,5 @@
 import externalLinkIcon from '@/assets/svg/external-link.svg';
-import Image from '@/features/image/image';
+import Image from '@/shared/image/image';
 
 export const ExternalLinkIcon = () => {
   return <Image src={externalLinkIcon} alt="external link icon" />;
