+++ src/icons/github.tsx
@@ -1,5 +1,5 @@
 import github from '@/assets/svg/github.svg';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 export const GithubLogo = () => {
   return <Image src={github} alt=" github icon" />;
