+++ src/icons/html-icon.tsx
@@ -1,5 +1,5 @@
 import html from '@/assets/svg/html.svg';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 export const HtmlIcon = () => {
   return <Image src={html} alt="html icon" />;
