+++ src/icons/javascript-icon.tsx
@@ -1,5 +1,5 @@
 import javascript from '@/assets/icons/js-secondary.webp';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 export const JavascriptIcon = () => {
   return <Image src={javascript} alt="JavaScript icon" />;
