+++ src/icons/nodejs-icon.tsx
@@ -1,5 +1,5 @@
 import nodeJS from '@/assets/svg/nodeJS.svg';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 export const NodeJsIcon = () => {
   return <Image src={nodeJS} alt="nodeJS icon" />;
