+++ src/icons/open-source-philosophy-icon.tsx
@@ -1,5 +1,5 @@
 import openSourcePhilosophyIcon from '@/assets/svg/openSourcePhilosophyIcon.svg';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 export const OpenSourcePhilosophyIcon = () => {
   return <Image src={openSourcePhilosophyIcon} alt="Open Source Philosophy Icon" />;
