+++ src/icons/open-to-everyone-icon.tsx
@@ -1,5 +1,5 @@
 import openToEveryoneIcon from '@/assets/svg/openToEveryoneIcon.svg';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 export const OpenToEveryoneIcon = () => {
   return <Image src={openToEveryoneIcon} alt="Open To Everyone Icon" />;
