+++ src/icons/react-icon.tsx
@@ -1,5 +1,5 @@
 import react from '@/assets/svg/react.svg';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 export const ReactIcon = () => {
   return <Image src={react} alt="react icon" />;
