+++ src/icons/rs-banner.tsx
@@ -1,5 +1,5 @@
 import banner from '@/assets/svg/RsBanner.svg';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 export const RsBanner = () => {
   return <Image data-testid="rs-banner" src={banner} alt="RsBanner Icon" />;
