+++ src/icons/teach-It-forward-icon.tsx
@@ -1,5 +1,5 @@
 import teachItForwardIcon from '@/assets/svg/teachItForwardIcon.svg';
-import Image from '@/features/image';
+import Image from '@/shared/image';
 
 export const TeachItForwardIcon = () => {
   return <Image src={teachItForwardIcon} alt="Teach It Forward Icon" />;
