+++ src/shared/image/index.ts
