+++ src/shared/image/types.ts
